Abstract HeaderMap to be used elsewhere
