+++ src/sign.rs
@@ -11,44 +11,9 @@ use serde_bytes::ByteBuf;
 use serde_cbor::Error as CborError;
 use serde_cbor::Value as CborValue;
 use serde_repr::{Deserialize_repr, Serialize_repr};
-use std::collections::BTreeMap;
 
 use crate::error::COSEError;
-
-#[derive(Clone, Debug, Serialize, Deserialize, Default)]
-/// Implementation of header_map, with CborValue keys and CborValue values.
-pub struct HeaderMap(
-    #[serde(deserialize_with = "::serde_with::rust::maps_duplicate_key_is_error::deserialize")]
-    BTreeMap<CborValue, CborValue>,
-);
-
-impl HeaderMap {
-    /// Creates an empty HeaderMap
-    pub fn new() -> Self {
-        HeaderMap(BTreeMap::new())
-    }
-
-    /// Inserts an element into HeaderMap. Both key and value are CborValue.
-    /// If key already has a value, that value is returned.
-    pub fn insert(&mut self, key: CborValue, value: CborValue) -> Option<CborValue> {
-        self.0.insert(key, value)
-    }
-
-    /// Returns the element at key.
-    pub fn get(&self, key: &CborValue) -> Option<&CborValue> {
-        self.0.get(key)
-    }
-
-    /// Returns true if HeaderMap has no elements, false otherwise.
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-
-    /// Parses a slice of bytes into a HeaderMap, if possible.
-    pub fn from_bytes(header_map: &[u8]) -> Result<Self, CborError> {
-        serde_cbor::from_slice(&header_map)
-    }
-}
+use crate::header_map::{map_to_empty_or_serialized, HeaderMap};
 
 /// Values from https://tools.ietf.org/html/rfc8152#section-8.1
 #[derive(Debug, Copy, Clone, Serialize_repr, Deserialize_repr)]
@@ -130,14 +95,6 @@ pub struct SigStructure(
     /// payload : bstr
     ByteBuf,
 );
-
-fn map_to_empty_or_serialized(map: &HeaderMap) -> Result<Vec<u8>, CborError> {
-    if map.is_empty() {
-        Ok(vec![])
-    } else {
-        Ok(serde_cbor::to_vec(map)?)
-    }
-}
 
 impl SigStructure {
     /// Takes the protected field of the COSE_Sign object and a raw slice of bytes as payload and creates a
